pyproject.toml
[tool.poetry]
name = "free-smiley-dealer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "1.3.1"
pymongo = "3.10.1"
dblpy = "*"
environs = "^7.3.1"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
